små fix
